python-fastapi/requirements.txt
Nuitka==2.2.1
fastapi==0.111.0
tortoise-orm[aiosqlite]==0.20.1
asyncpg==0.29.0
